Add option to disable brackets around table names in bcp calls
Previous behavior was to enclose all table names in brackets. E.g.,
"dbo.customers" -> "[dbo].[customers]". However, this appears to cause
failure when attempting to import data from a linux system to an
Azure Synapse database. This change adds a flag to disable the table
name "quoting" while maintaining it as a default.